Basic Setup
Created an empty control node scene for the Dialog System; i plan to use this later, but if i don't it WILL be deleted.
Otherwise, set up export presets so we can test builds in browser; we need a designated main scene for this, but that comes later.
Lastly, set up some sensible (in my opinion) defaults for project settings