Added destructor
Destructor allows items to destroy themselves.
Would rather write it so that the destructor itself can use item_node_name.queue_free() so that it only needs to be written once.

Also need to change naming scheme for items in draw to
Item_name_canvas
item_name
so the names aren't allover the place