Creation of Barter System Skeleton, and Creation of Generic Character
fuck off github desktop, i'll make the summaries as long as i please.
Anyways, created the generic character: this should be the instantiated the character whenever a new character is generated.
Also worked on the barter system; for now it's just a very barebones concept, but should be expanded eventually maybe.